+++ R/plot_decontPro.R
@@ -60,7 +60,7 @@ plotDensity <- function(counts,
     # Plot
     p1 <- ggplot2::ggplot(
       df.m,
-      ggplot2::aes_string("value", fill = "variable")
+      ggplot2::aes(.data$value, fill = .data$variable)
     ) +
       ggplot2::geom_density(alpha = 0.7) +
       ggplot2::scale_x_continuous(
@@ -182,10 +182,10 @@ plotBoxByCluster <- function(counts,
     # Plot
     p1 <- ggplot2::ggplot(
       df.m,
-      ggplot2::aes_string(
-        x = "cell_type",
-        y = "value",
-        fill = "variable"
+      ggplot2::aes(
+        x = .data$cell_type,
+        y = .data$value,
+        fill = .data$variable
       )
     ) +
       ggplot2::geom_boxplot(